:recycle: Move schema to models
